requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-dotenv==1.0.0
drf-yasg==1.21.7 